refactor: unique identifier transform
