+++ transforms/graphql_tag/src/lib.rs
@@ -2,20 +2,21 @@
 use std::collections::HashMap;
 
 // libs
-use swc_common::{comments::Comments, BytePos, SourceFile};
+use swc_common::comments::Comments;
 use swc_ecma_ast::*;
-use swc_ecma_parser::parse_file_as_expr;
 use swc_ecma_visit::{VisitMut, VisitMutWith};
 
 // modules
 pub mod parser;
 pub mod structs;
+mod utils;
 
 // helpers
 use parser::utils::strip_ignored_characters;
 
 // structs
 use structs::{GraphQLTagConfig, TransformVisitor};
+use utils::add_unique_fn_to_program;
 
 impl<C> TransformVisitor<C>
 where
@@ -41,76 +42,7 @@ where
         self.active_gql_tag_identifiers.clear();
 
         if self.unique_fn_used {
-            let source: String = "(definitions) => {
-            const names = {};
-            return definitions.filter(definition => {
-              if (definition.kind !== 'FragmentDefinition') {
-                return true;
-              }
-              const name = definition.name.value;
-              if (names[name]) {
-                return false;
-              } else {
-                names[name] = true;
-                return true;
-              }
-            });
-          }"
-            .into();
-
-            let source_file = SourceFile::new(
-                swc_common::FileName::Anon,
-                false,
-                swc_common::FileName::Anon,
-                source,
-                BytePos(1),
-            );
-
-            let expr_result = parse_file_as_expr(
-                &source_file,
-                Default::default(),
-                Default::default(),
-                Default::default(),
-                &mut vec![],
-            )
-            .expect("failed to create unique function");
-
-            match node {
-                Program::Module(program) => program.body.insert(
-                    1,
-                    ModuleItem::Stmt(Stmt::Decl(Decl::Var(Box::new(VarDecl {
-                        span: program.span,
-                        kind: VarDeclKind::Const,
-                        declare: false,
-                        decls: vec![VarDeclarator {
-                            span: program.span,
-                            name: Pat::Ident(
-                                Ident::new(self.config.unique_fn_name.clone().into(), program.span)
-                                    .into(),
-                            ),
-                            init: Some(expr_result),
-                            definite: true,
-                        }],
-                    })))),
-                ),
-                Program::Script(program) => program.body.insert(
-                    1,
-                    Stmt::Decl(Decl::Var(Box::new(VarDecl {
-                        span: program.span,
-                        kind: VarDeclKind::Const,
-                        declare: false,
-                        decls: vec![VarDeclarator {
-                            span: program.span,
-                            name: Pat::Ident(
-                                Ident::new(self.config.unique_fn_name.clone().into(), program.span)
-                                    .into(),
-                            ),
-                            init: Some(expr_result),
-                            definite: true,
-                        }],
-                    }))),
-                ),
-            }
+            add_unique_fn_to_program(node, self.config.unique_fn_name.clone())
         }
     }
