+++ transforms/unique_identifier/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+authors = ["Rishabh Chawla <rishabh31121999@gmail.com>"]
+name = "unique_identifier"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+apollo-parser = "0.2.10"
+swc_common = { version = "0.33.8", features = ["concurrent"] }
+swc_ecma_ast = "0.110.9"
+swc_ecma_visit = "0.96.9"
+swc_ecma_parser = "0.141.33"
+swc_core = { version = "0.86.55", features = ["ecma_plugin_transform"] }
+regex = "1"
+serde = "1.0.193"
+serde_json = "1.0.108"
